+++ RELEASE-NOTES.md
@@ -81,6 +81,7 @@ The Axom project release numbers follow [Semantic Versioning](http://semver.org/
 - Updated built-in TPL `fmt` to master branch snapshot, March 26, 2021.
 - Inlet: SphinxWriter will now print only one element schema per container instead of
   printing the same schema for each element in the container
+- Inlet: `Writer`s are passed directly to `Inlet::write` instead of being registered
 
 ### Fixed
 - Updated to new BLT version that does not fail when ClangFormat returns an empty